+++ lib/flutter_earth_globe.dart
@@ -1,27 +1,42 @@
 library flutter_earth_globe;
 
+import 'globe_coordinates.dart';
 import 'rotating_globe.dart';
 import 'package:flutter/material.dart';
 
-import 'rotating_globe_controller.dart';
+import 'flutter_earth_globe_controller.dart';
 
 /// This is the main widget of the package. It is a sphere that can be rotated and animated.
 class FlutterEarthGlobe extends StatefulWidget {
   final double radius;
   final Alignment alignment;
-  final RotatingGlobeController controller;
+  final FlutterEarthGlobeController controller;
+  final void Function(double zoom)? onZoomChanged;
 
-  /// The [FlutterEarthGlobe] widget represents a 3D sphere that simulates the Earth globe. It can be rotated and animated using the provided [RotatingGlobeController].
+  final void Function(GlobeCoordinates? coordinates)? onHover;
+
+  final void Function(GlobeCoordinates? coordinates)? onTap;
+
+  /// The [FlutterEarthGlobe] widget represents a 3D sphere that simulates the Earth globe. It can be rotated and animated using the provided [FlutterEarthGlobeController].
   ///
   /// The [radius] parameter specifies the radius of the sphere.
   /// The [alignment] parameter specifies the alignment of the sphere within its parent widget. By default, it is centered.
   /// The [controller] parameter is responsible for controlling the rotation and animation of the globe.
+  /// The [onZoomChanged] parameter is a callback function that is called when the zoom level of the globe changes.
+  /// The [onHover] parameter is a callback function that is called when the user hovers over a point on the globe.
+  /// The [onTap] parameter is a callback function that is called when the user taps on a point on the globe.
   ///
   /// Example usage:
   /// ```dart
   /// FlutterEarthGlobe(
   ///  radius: 200,
-  /// controller: RotatingGlobeController(),
+  /// controller: FlutterEarthGlobeController(),
+  /// onZoomChanged: (zoom) {
+  /// print('Zoom level: $zoom');
+  /// },
+  /// onHover: (coordinates) {
+  /// print('Hovering over coordinates: $coordinates');
+  /// },
   /// )
   /// ```
   const FlutterEarthGlobe({
@@ -29,6 +44,9 @@ class FlutterEarthGlobe extends StatefulWidget {
     required this.radius,
     required this.controller,
     this.alignment = Alignment.center,
+    this.onZoomChanged,
+    this.onHover,
+    this.onTap,
   }) : super(key: key);
 
   @override
@@ -38,10 +56,13 @@ class FlutterEarthGlobe extends StatefulWidget {
 class _FlutterEarthGlobeState extends State<FlutterEarthGlobe> {
   @override
   Widget build(BuildContext context) {
-    return Sphere(
+    return RotatingGlobe(
       controller: widget.controller,
       radius: widget.radius,
       alignment: widget.alignment,
+      onZoomChanged: widget.onZoomChanged,
+      onHover: widget.onHover,
+      onTap: widget.onTap,
     );
   }
 }