+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+- Added more listeners and callbacks
+- Improved overall usage of controller
+  
 ## 1.0.1
 - Minor Readme fixes